requirements.txt
Django>=4.0,<5.0
PyMuPDF
pdfplumber
pandas
openpyxl
pdf2image
pytesseract
python-docx
gunicorn
